Cert expiry monitor script (#40)
Simple script and Jenkins job to monitor certs for impending expiries.
The script alerts on the #devops Slack channel when a cert is less than
four weeks away from its expiry date.  And the Jenkins job for this will
be configured to run once a day.

The Jenkins job is currently configured to monitor registry and vault.