Reverse-engineered edge-cloud docker base image (#55)
* Reverse-engineered edge-cloud docker base image

The current edge-cloud docker image
("mobiledgex/openstack-kubectl-gcloud-az") does not have a Dockerfile
anywhere.  All there is is an image which we use everywhere but cannot
rebuild.  Also, there are a lot of config files in the root directory
which are not under version control.

This Dockerfile is reverse-engineered from the
mobiledgex/openstack-kubectl-gcloud-az image and contains all the
packages it had, as well as the full contents of the root directory.

There are a lot of stale files in the root directory which will be
cleaned up in a subsequent commit.

* Remove logs from original push