EDGECLOUD-540 CICD for edge-cloud-infra (#77)
Added an EDGE_CLOUD_VERSION Makefile variable which formalises the dependency between the edge-cloud-infra repo and edge-cloud. For master, edge-cloud-infra depends on the "master" branch on edge-cloud. But when we branch off a release, we tag edge-cloud and infra and set the EDGE_CLOUD_VERSION to the release tag of edge-cloud.

There are two new Make targets:
- edge-cloud-version-check: Mandatory check before a build to make sure the edge-cloud repo is
  in the right branch/tag
- edge-cloud-version-set: Primarily for Jenkins CICD to automatically set the edge-cloud repo to
  the branch/tag required by the build