Switch to new statefulset etcd
The changes here are to manage the new statefulset etcd clusters _after_
the migration from the old etcd-operator.  The migration playbooks
themselves are in the "etcd-oper-migration" branch along with a
"ansible/ETCD-MIGRATION.txt" guide, but they are now essentially
obsolete as the main medemo, staging, and dev environments have already
been migrated.