DME cert renewal
DME will now retrieve its own certs from vault using the "vault-cert"
script.

The same script is used as a liveness probe that'll fail if there is a
newer cert available in vault.  The liveness check happens every 8
hours, and if three successive checks fail, the container will be
restarted.