EDGECLOUD-2302 Add GPU support to mobiledgex base image (#714)
NVIDIA driver 440 is not available for the Ubuntu 16.04 which our base
image is built on.  Had to build a custom package manually using
"checkinstall" and NVIDIA's run script:
https://download.nvidia.com/XFree86/Linux-x86_64/440.64/NVIDIA-Linux-x86_64-440.64.run

The custom nvidia-440 package is available in Artifactory's APT repo:
https://artifactory.mobiledgex.net/artifactory/webapp/#/artifacts/browse/tree/General/packages/pool

Have also mirrored the NVIDIA container toolkit repo in apt.mobiledgex.net.