Support for deployment of production console builds (#808)
The `console_prod` parameter controls whether the development or
production build of the console gets deployed in a setup.  Currently,
main and staging are configured to have the production version deployed,
and QA and dev will use the development version.

This can be overridden at deployment time by passing the `console_prod`
parameter as an argument to `deploy.sh`. For instance, to deploy the
production version of the console in QA:

```
./deploy.sh -e console_prod=yes qa console
```

The actual builds of the dev and prod images happen in the edge-cloud-ui
repo.