EDGECLOUD-3097 TEF cloudlet unable to download from artifactory
Nginx buffers upstream Artifactory downloads, but if it has a limit of
1G for the temporary buffer file that it creates.  If the download
connection is fast, it will not hit this buffer limit, but with slower
connections, nginx waits for the client and stop downloading from
Artifactory, and Artifactory interprets that as a socket timeout and
tears down the connection.  Once the client downloads 1GB, nginx tries
to resume the download from Artifactory and fails.