EDGECLOUD-3500 Creation of docker compose based App Instance fails (#972)
Docker Compose is packaged using PyInstaller which unpacks its python
runtime in /tmp and executes it there.  However, CIS hardening and
general security recommendations require /tmp to be mounted with the
"noexec" option which disallows executables within that mount point.

Enabling "exec" in /tmp is not recommended, and setting TMPDIR globally
affects all apps which would run on the VM. Hence, wrapping
docker-compose in a script which sets up TMPDIR just for it.  The actual
docker-compose binary is installed in /usr/local/lib.