EDGECLOUD-3735 Get rid of python-pip from docker base image
Pip pulls in a ton of dependencies, including the entire build-essential
metapackage which includes gcc and perl. This bloats up the image by
about 170Mb and also increase the attack surface for vulnerabilities.

Fixing this by building pip and installing the python modules in a
separate build stage and copying just the python files into the final
image.