Fix shard limit issue in Elasticsearch (#1223)
With the number of setups we have currently, we have used up the shard
limit in ElasticSearch. We have daily indexes for events, and each index
has ten shards.  The upper limit for the cluster is set to 2000 shards,
which we have hit already.

This is mitigation rather than a complete fix. We probably need to
switch from daily indexes to monthly ones for events. The number of
indexes/shards blows up memory usage requirements, and the amount of
data we have in each daily index is pretty small right now---about 20MB
per day in the QA setup.