Make sure apt sources directory is present in base image
If the baseimage is new enough (4.x), the `/etc/apt/sources.list.d`
directory would be missing, having been deleted during baseimage
creation.  However, the new chef recipes use this directory to manage
apt sources.

Have also changed the part of the baseimage build (`setup.sh`) which was
deleting this directory. There is no need to build an updated baseimage
just for this change, though. It is sufficient to include this change in
any subsequent baseimage builds.