Flag to control actual staging deployment (#1720)
Right now, on every console release, a Github action attempts to
deploy the console to the staging setup. However, there are times when
the staging setup is locked down, and this ends up triggering a false
failure in the Github action.

With this change, we can control the actual deployment of the console
with a flag (manually configued in Jenkins), and if the flag is not set,
the Jenkins job will succeed without actually doing anything.