Monitor Jaeger metrics (#1783)
With this change, jaeger metrics will be logged to our internal
InfluxDB. The primary metric we are interested in the number of dropped
spans, which can happen if the backend data (like ElasticSearch) is
slow. Will add alerts for this as a separate commit after monitoring
the data for a few days.