Pod autoscaling for controller and DME
Support for autoscaling controller and DME pods depends on CPU and
memory usage. Defaults are set up to scale when CPU or memory usage
go over 50% of the configured initial request. (AKS seems to have
trouble with averageUtilization percentage triggers for memory; using
calculated averageValue for that instead.)

The default maximum replicas for all services remains at 1 right now,
which means this change will have no impact on any setup except for QA
which has controller replicas set to 2.  When we are ready for
controller and DME to scale, will set the default (maximum) replica
count to 3 and let the autoscaler decide the number of pods to run.