Pod autoscaling for controller and DME (#1796)
Support for autoscaling controller and DME pods depends on CPU and
memory usage. Defaults are set up to scale when CPU or memory usage
go over 50% of the configured initial request. (AKS seems to have
trouble with averageUtilization percentage triggers for memory; using
calculated averageValue for that instead.)

The minimum number of replicas is set to 2 and the max 3 for all setups.