Support for short-lived kubeconfigs for CI/CD
Teleport supports generating short-lived kubeconfigs which can be used
for CI/CD. The kubeconfigs are generated on the Teleport host and are
stored in vault, where ansible picks them up during deployments.

The kubeconfigs are generated with a TTL of 49 hours, and a new
kubeconfig is generated and published to vault every 24 hours.

Reference: https://goteleport.com/docs/kubernetes-access/guides/cicd/