EDGECLOUD-5326 Automatic scanning of developer docker (#1902)
VM to scan developer docker images.

This infrastructure can be dismantled once we replace Gitlab with
Harbor. Harbor can be configured to scan docker images automatically,
but to get the equivalent functionality in Gitlab, we will need to set
up a CI job for this in every project (i.e., every developer org) in
Gitlab. That'd be a nightmare to configure and maintain, and especially
pointless as this will all go away once we get Harbor.

This external test setup is configured to run image scans on demand, and
there are jobs in Roma which track and scan all developer images
periodically.
https://roma.mobiledgex.net/inventory/appimage