Implement minimum version checks for teleport (#1936)
Teleport pulls older versions of their packages from their APT repo when
a newer version is released. This breaks deployments unless we have the
exact latest version number listed in our playbooks.

Switch to ensuring a minimum version of teleport, and if the installed
version is older than that, upgrade to the latest version.